Removed unnecessary images and corrected Herbalism icon
The Herbalism icon was still showing the old version. Turns out it was because the ingame icon from the game files is named 'Gathering' instead. Renamed it to Herbalism so that it works with the way this calculator is scripted. Removed a bunch of images that aren't being utilized at the moment.